Fix for the issue 48
It is a fix for the issue when user tries to open a file without
extension throw an ArrayIndexoutofboundry error.
If user tries to open file then application try to determine the file
extension even if it does not have any extension, which is causing this
error. This change will check if file extension exist or not and select
logic appropriately.
Related issue: https://github.com/nccgroup/ncccodenavi/issues/48